refactoring redux use rtk
