+++ src/IndexPage.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchArticleList } from './actions';
+import { useDispatch } from 'react-redux';
+import { fetchArticleList } from './slices';
 import './Nav.scss';
 import IdeaShow from './IdeaShow';
 import IndexContentContainer from './IndexContentContainer';
@@ -9,12 +9,10 @@ function IndexPage(){
     const dispatch = useDispatch();
 
     useEffect(() => {
-        console.log('mount');
     }, []);
 
     useEffect(() => {
         dispatch(fetchArticleList());
-        console.log('triggered');
     }, [dispatch]);
 
     return (
