--- src/actions/actionTypes.js
@@ -1,2 +1,0 @@
-export const FETCH_ARTICLE_LIST = 'FETCH_ARTICLE_LIST';
-export const FETCH_ARTICLE_COMMENTS = 'FETCH_ARTICLE_COMMENTS';
