--- src/actions/index.js
@@ -1,2 +1,0 @@
-export * from './actionTypes.js';
-export * from './articleActions.js';
