--- src/reducers/index.js
@@ -1,5 +1,0 @@
-import { articleList } from './articleListReducer';
-
-export default {
-    articleList,
-};
