+++ src/slices/index.js
@@ -1,0 +1,7 @@
+import articleList from './articleListSlice';
+export * from './articleListSlice';
+
+
+export default {
+    articleList
+};
